requirements.txt
selenium
pandas
python-dateutil
webdriver-manager